% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevarying_error.R
\name{error_measures}
\alias{error_measures}
\title{Evaluate prediction performance}
\usage{
error_measures(qx, pred_qx, wx = NULL)
}
\arguments{
\item{qx}{Matrix of observed mortality rates}

\item{pred_qx}{List of matrices (one matrix for each method) of predicted qx with rows corresponding to the ages of the observed qx. Each column should correspond to a different cohort.}

\item{wx}{Matrix or vector of weights for the weighted prediction error measures. If vector, vector should be the starting cohort size for each cohort.}
}
\value{
A vector of prediction accuracy measures.
}
\description{
Evaluate prediction performance of any time-varying model using root mean squared error, mean absolute error, weighted root mean squared error and weighted mean absolute error.
}
\examples{
qx <- matrix(runif(15, 0.01, 0.3), nrow = 5, ncol = 3)
pred1 <- qx + matrix(rnorm(15, 0, 0.02), nrow = 5)
pred2 <- qx + matrix(rnorm(15, 0, 0.03), nrow = 5)

# Starting cohort sizes
start_weights <- c(1000, 1500, 2000)

# Evaluate
error_measures(
  qx = qx,
  pred_qx = list(Method_A = pred1, Method_B = pred2),
  wx = start_weights
)

}
