% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getqx.R
\name{get_qx}
\alias{get_qx}
\title{Get predictions from MortalityLaws models}
\usage{
get_qx(x, qx, law, pred_ages)
}
\arguments{
\item{x}{Vector of ages at the beginning of the age interval}

\item{qx}{Vector of observed mortality rates}

\item{law}{The law to get predictions from - see MortalityLaws package}

\item{pred_ages}{ages at which to predict qx}
}
\value{
predicted qx from chosen law
}
\description{
Extract predicted qx from models fit using the MortalityLaws package
}
\examples{
qx <- seq(0.1,0.5,0.01)
get_qx(60:100, qx, "HP2", 80:105)

}
