% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdstlt.R
\name{plot.dstlt}
\alias{plot.dstlt}
\title{Plot the Dyanmic Smooth Threshold Life Table}
\usage{
\method{plot}{dstlt}(x, main = "", sub = "", ...)
}
\arguments{
\item{x}{a model object of class dstlt}

\item{main}{title of plot}

\item{sub}{subtitle of plot}

\item{...}{Additional arguments passed to methods}
}
\description{
Plot fitted DSTLT lines for the cohorts used to fit the model along with corresponding observed qx
}
\examples{
dstlt.mod<-dstlt(cbind(60:100,60:100,60:100),cbind(seq(0.1,0.5,0.01),seq(0.1,0.5,0.01),seq(0.1,0.5,0.01)),endN=95)
plot(dstlt.mod)

}
