% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotstlt.R
\name{plot.stlt}
\alias{plot.stlt}
\title{Plot the Smooth Threshold Life Table}
\usage{
\method{plot}{stlt}(x, main = "", sub = "", ...)
}
\arguments{
\item{x}{An stlt model object}

\item{main}{title of plot}

\item{sub}{subtitle of plot}

\item{...}{Additional arguments passed to methods}
}
\description{
Plots the fitted threshold life table line and corresponding observed qx used for fitting the model.
}
\examples{
stlt.mod<-stlt(60:100,seq(0.1,0.5,0.01))
plot(stlt.mod)

}
