% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictstlt.R
\name{predict.stlt}
\alias{predict.stlt}
\title{Predict using a fitted Smooth Threshold Life Table}
\usage{
\method{predict}{stlt}(object, newdata, ...)
}
\arguments{
\item{object}{an STLT model object}

\item{newdata}{Vector of ages for which qx are to be predicted}

\item{...}{Additional arguments passed to methods}
}
\value{
Vector of predicted qx under the fitted STLT model
}
\description{
For a given STLT fit, this function computes the qx associated with any particular age
}
\examples{
stlt.mod<-stlt(60:100,seq(0.1,0.5,0.01))
predict(stlt.mod,65:90)

}
