% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparemethods.R
\name{compare_methods}
\alias{compare_methods}
\title{Fit the Smooth Threshold Life Table}
\usage{
compare_methods(qx, pred_qx, wx = NULL)
}
\arguments{
\item{qx}{Vector of observed mortality rates; ages should be integer values and sequential}

\item{pred_qx}{A matrix of predicted qx, corresponding to the ages of the observed qx. Each column should correspond to a different method used to predict qx.}

\item{wx}{Vector of weights for the weighted prediction error measures. Usually the cohort size at each x. Can be a scalar indicating the cohort size at the first age - subsequent weights will then be calculated using the observed qx.}
}
\value{
A list of prediction accuracy measures
}
\description{
This package fits the Smooth Treshold Life Table and Dyanmic Smooth Threshold Life Table, as in Huang et al., (2019). Fitted and predicted qx as well as their plots are provided. No right censoring is applied and all possible ages are considered when estimating the threshold age N.
}
\examples{
qx <- c(0.1, 0.2, 0.15, 0.3)
pred_qx <- matrix(c(0.12, 0.11,0.21, 0.19,0.13, 0.18,0.31, 0.29), ncol = 2)
colnames(pred_qx) <- c("Method_A", "Method_B")
compare_methods(qx, pred_qx, wx = 10000)

}
