% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstlt.R
\name{dstlt}
\alias{dstlt}
\title{Fit the Dyanmic Smooth Threshold Life Table}
\usage{
dstlt(ages, qxs, startN = 80, endN = 105, censorAge = NULL, hessian = FALSE)
}
\arguments{
\item{ages}{Matrix of ages that are used to fit the DSTLT. Each column should represent one cohort. There should not be missing data for any ages between the smallest and largest age.}

\item{qxs}{Matrix of mortality rates that are used to fit the DSTLT, corresponding to the matrix of ages. Each column should represent one cohort.The qx matrix need not be equal to one for any element; right censoring is accounted for in the likelihood function. Where the highest observed age at death is different across cohorts, elements in the qx matrix should be filled with NA for ages past the highest observed age for each cohort.}

\item{startN}{The first potential threshold age N tested}

\item{endN}{The last potential threshold age N tested}

\item{censorAge}{The age at which to right-censor observations}

\item{hessian}{set to TRUE for standard errors}
}
\description{
This package fits the Smooth Threshold Life Table and Dyanmic Smooth Threshold Life Table, as in Huang et al., (2019). Fitted and predicted qx as well as their plots are provided. No right censoring is applied and all possible ages are considered when estimating the threshold age N.
}
\examples{
dstlt(cbind(60:100,60:100,60:100),cbind(seq(0.11,0.51,0.01),seq(0.105,0.505,0.01),seq(0.1,0.5,0.01)),endN=95)

}
