% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstlt.R
\name{dstlt}
\alias{dstlt}
\title{Fit the Dyanmic Smooth Threshold Life Table}
\usage{
dstlt(ages, qxs, startN = 80, endN = 105, censorAge = NULL)
}
\arguments{
\item{ages}{Matrix of ages that are used to fit the DSTLT. Each column should represent one cohort.}

\item{qxs}{Matrix of mortality rates that are used to fit the DSTLT. Each column should represent one cohort.}

\item{startN}{The first potential threshold age N tested}

\item{endN}{The last potential threshold age N tested}

\item{censorAge}{The age at which to right-censor observations}
}
\description{
This package fits the Smooth Threshold Life Table and Dyanmic Smooth Threshold Life Table, as in Huang et al., (2019). Fitted and predicted qx as well as their plots are provided. No right censoring is applied and all possible ages are considered when estimating the threshold age N.
}
\examples{
dstlt(cbind(60:100,60:100,60:100),cbind(seq(0.1,0.5,0.01),seq(0.1,0.5,0.01),seq(0.1,0.5,0.01))

}
