% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlt.R
\name{stlt}
\alias{stlt}
\title{Fit the Smooth Threshold Life Table}
\usage{
stlt(
  ages,
  qx,
  startN = 0,
  endN = 200,
  censorAge = NULL,
  hessian = FALSE,
  radix = 50000
)
}
\arguments{
\item{ages}{Vector of ages that are used to fit the STLT. There should not be missing data for any ages between the smallest and largest age.}

\item{qx}{Vector of mortality rates that are used to fit the STLT, corresponding to the vector of ages. The qx vector need not be equal to one for any element; right censoring is accounted for in the likelihood function.}

\item{startN}{The first potential threshold age N tested}

\item{endN}{The last potential threshold age N tested}

\item{censorAge}{The age at which to right-censor observations}

\item{hessian}{Set to TRUE for standard errors}

\item{radix}{The radix for the hypothetical cohort; affects standard errors}
}
\value{
A model object of class "stlt"
}
\description{
This package fits the Smooth Threshold Life Table and Dyanmic Smooth Threshold Life Table, as in Huang et al., (2020). Fitted and predicted qx as well as their plots are provided. No right censoring is applied and all possible ages are considered when estimating the threshold age N.
}
\examples{
stlt(60:100,seq(0.1,0.5,0.01))

}
